change int

+++ Schema/cinema.sql
@@ -20,7 +20,7 @@ CREATE TABLE customer (
 );
 
 CREATE TABLE  film_booking (
-	booking_price_id DECIMAL, 
+	booking_price_id INT, 
 	fk_customer_id INT,
 	movie_restrictions VARCHAR(70),
 	movie_avaliabilty DATETIME NOT NULL,
